requirements.txt
pipecat-ai[openai,silero,deepgram,google,webrtc]
fastapi
uvicorn
dotenv
# Core dependencies
pydantic>=2.0.0
# Google Cloud dependencies
google-generativeai>=0.3.0
supabase
google-genai
python-jose
python-dotenv